Update merged components & address feedback.
