+++ dev/traffic_portal/run.sh
@@ -35,5 +35,5 @@ if [[ "$(id -u)" != "$uid" ]]; then
 	exec su "$user" -- "$0"
 fi
 
-npm i --save-dev
+npm i
 ./node_modules/.bin/grunt
