+++ experimental/traffic-portal/nightwatch/page_objects/cacheGroups/cacheGroupsTable.ts
@@ -13,8 +13,7 @@
  */
 
 import { EnhancedPageObject, EnhancedSectionInstance, NightwatchAPI } from "nightwatch";
-
-import { TABLE_COMMANDS, TableSectionCommands } from "../globals/tables";
+import { TABLE_COMMANDS, TableSectionCommands } from "../../globals/tables";
 
 /**
  * Defines the Cache Group table commands
