+++ experimental/traffic-portal/src/app/core/servers/phys-loc/detail/phys-loc-detail.component.html
@@ -37,11 +37,11 @@ limitations under the License.
 				<input matInput type="text" name="state" required [(ngModel)]="physLocation.state" />
 			</mat-form-field>
 			<mat-form-field>
-				<mat-label>Zip</mat-label>
-				<input matInput type="text" name="zip" required [(ngModel)]="physLocation.zip" />
+				<mat-label>Postal Code</mat-label>
+				<input matInput type="text" name="postalCode" required [(ngModel)]="physLocation.zip" />
 			</mat-form-field>
 			<mat-form-field>
-				<mat-label>Poc</mat-label>
+				<mat-label>Person of Contact</mat-label>
 				<input matInput type="text" name="poc" [(ngModel)]="physLocation.poc" />
 			</mat-form-field>
 			<mat-form-field>
